cloud_function/requirements.txt
postgresql-interface==1.0.2
python-dotenv==0.14.0
google-cloud-storage==1.42.3